feat(booking): update airline-initiated change types and move to separate folder
